made mock_env available to end to end tests as well
