+++ tests/unit_test/conftest.py
@@ -1,6 +1,7 @@
 import logging
 from unittest.mock import MagicMock, patch
-from typing import Iterable, Iterator
+from typing import Iterator
+
 import boto3
 import pytest
 
@@ -10,13 +11,6 @@ def setup_logging():
     logging.basicConfig(level='INFO')
     for name in ['tests', 'dags', 'data_pipeline']:
         logging.getLogger(name).setLevel('DEBUG')
-
-
-@pytest.fixture()
-def mock_env() -> Iterable[dict]:
-    env_dict: dict = {}
-    with patch('os.environ', env_dict):
-        yield env_dict
 
 
 @pytest.fixture(name="mock_s3_client_function", autouse=True)
